+++ server_test.go
@@ -21,7 +21,7 @@ import (
 	"crypto/rand"
 	"github.com/loopholelabs/frisbee-go/pkg/metadata"
 	"github.com/loopholelabs/frisbee-go/pkg/packet"
-	"github.com/loopholelabs/polyglot-go"
+	"github.com/loopholelabs/polyglot"
 	"github.com/loopholelabs/testing/conn"
 	"github.com/loopholelabs/testing/conn/pair"
 	"github.com/rs/zerolog"